grepify/pyproject.toml
[project]
name = "grepify"
version = "0.1.0"
description = "RAG-powered search over subreddits"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "sentence-transformers>=3.0",
    "qdrant-client>=1.9",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["grepify"]
